Send back response when timeout
When there is a connection timeout currently the channel is closed. In this PR if the timeout is during the reception of a request or after completion of the reception of the request a response is sent back to inform the client. If it is the timeout when reading the request 408 Request Timeout is sent. If it is the timeout before writing a response a 500 internal server error is sent.

Resolve https://github.com/ballerina-platform/ballerina-lang/issues/3794