Add onClose method to the ConnectorListener
The onClose method for the WebSocketConnectorListener has been introduced in order to let the listener know when the connection closes. This method will be called only after the user is notified of a close message or an error message. Also note that this method will be called only after the futures of connection closure methods return.

Related issue: https://github.com/ballerina-platform/ballerina-lang/issues/10667